オブジェクトパネル開く
