+++ Assets/Shopping/Script/AssetDatabase.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 9a2a01ca574681843a97d36fd58d3552
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
