+++ Assets/Shopping/Script/AssetDatabase/ItemDatabase.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: ae8b7b192c337904a80b444c6bde4b3c
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
