+++ Assets/Shopping/Script/AssetDatabase/ItemDatabase.cs
@@ -1,0 +1,9 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+[CreateAssetMenu]
+public class ItemDatabase : ScriptableObject {
+    // AssetItemのリスト
+    public List<AssetItem> itemList = new List<AssetItem>();
+}
