--- Assets/Shopping/Script/AssetItem.cs
@@ -1,10 +1,0 @@
-using System;
-using UnityEngine;
-
-[Serializable]
-public class AssetItem {
-    // アイテムの外見
-    public GameObject item;
-    // アイテムの説明文
-    public string explainText;
-}
