--- Assets/Shopping/Script/AssetItem.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: a36bcc9ecdb1ced4c8d0820ef051165f
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
