--- Assets/Shopping/Script/ItemDatabase.asset
@@ -1,17 +1,0 @@
-%YAML 1.1
-%TAG !u! tag:unity3d.com,2011:
---- !u!114 &11400000
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 0}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 13c99fde17aeaef449e0635d31fd3f21, type: 3}
-  m_Name: ItemDatabase
-  m_EditorClassIdentifier: 
-  itemList:
-  - item: {fileID: 1031026121, guid: abc00000000013140153733274754931, type: 3}
-    explainText: test
