--- Assets/Shopping/Script/ItemDatabase.cs
@@ -1,9 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-[CreateAssetMenu]
-public class ItemDatabase : ScriptableObject {
-    // AssetItemのリスト
-    public List<AssetItem> itemList = new List<AssetItem>();
-}
