--- Assets/Shopping/Script/ItemDatabase.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 13c99fde17aeaef449e0635d31fd3f21
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
